Split off the RDRAND code
