+++ main.c
@@ -31,21 +31,7 @@
 #include <linux/random.h>
 
 #include "cpuid.h"
-
-// If rdrand fails, retry this many times
-static const int RETRY_COUNT = 10;
-
-// >1022 calls to 64-bit rdrand is guaranteed to include a reseed.
-// https://software.intel.com/en-us/articles/intel-digital-random-number-generator-drng-software-implementation-guide
-//   "The DRBG autonomously decides when it needs to be reseeded to refresh the
-//   random number pool in the buffer and is both unpredictable and transparent
-//   to the RDRAND caller. An upper bound of 511 128-bit samples will be
-//   generated per seed. That is, no more than 511*2=1022 sequential DRNG
-//   random numbers will be generated from the same seed value."
-#define RANDOM_COUNT 1024
-
-// The seed is 256 bits, so that's the amount of actual entropy we're adding
-static const int ENTROPY_INCREMENT = 256;
+#include "rand.h"
 
 // Add entropy at least this often, regardless of need (milliseconds)
 static const int MAX_SLEEP = 300000;
@@ -53,43 +39,13 @@ static const int MAX_SLEEP = 300000;
 // When woken up, fill until the pool has at least this much entropy
 static const int FILL_WATERMARK = 3072;
 
-static inline void fill_random_buf(uint64_t *buf)
+static void send_entropy(struct entropy (*get_entropy)(void), int fd)
 {
-    register unsigned int goodcalls = 0;
-    for (int i = 0; i < RANDOM_COUNT; i++)
-    {
-        register int counter = RETRY_COUNT;
-        register uint64_t val;
-        asm volatile ("1:\t"
-                      "dec %1\n\t"
-                      "je 2f\n\t"
-                      "rdrand %0\n\t"
-                      "jnc 1b\n\t"
-                      "2:\t"
-                      "adc $0,%2" : "=r" (val), "+g" (counter), "+g" (goodcalls) : : "cc");
-        buf[i] = val;
-    }
-    // Keeping a counter and checking later takes a branch out of the inner loop
-    if (goodcalls < RANDOM_COUNT) {
-        error(EXIT_FAILURE, 0,
-              "Error: rdrand failed %d times\n", RANDOM_COUNT - goodcalls);
-    }
-}
-
-static void send_entropy(int fd)
-{
-    struct {
-        int ent_count;
-        int size;
-        uint64_t buf[RANDOM_COUNT];
-    } entropy;
-    entropy.ent_count = ENTROPY_INCREMENT;
-    entropy.size = RANDOM_COUNT * sizeof(uint64_t);
-    fill_random_buf(entropy.buf);
+    struct entropy entropy = get_entropy();
     if (ioctl(fd, RNDADDENTROPY, &entropy) != 0) {
         perror("failed to add entropy");
     }
-    memset(entropy.buf, 0, RANDOM_COUNT * sizeof(uint64_t));
+    memset(entropy.buf, 0, BUFFER_SIZE * sizeof(uint64_t));
     // Tell gcc that buf is used, so it doesn't optimize away the memset
     asm ("" : : "m" (entropy.buf[0]) : "memory" );
 }
@@ -99,6 +55,8 @@ int main(int argc, char **argv)
     int random_fd, urandom_fd;
     int ent_count;
     struct pollfd pfd;
+
+    struct entropy (*get_entropy)(void) = get_entropy_rdrand;
 
     if (!has_rdrand()) {
         error(EXIT_FAILURE, 0, "This CPU does not support RDRAND");
@@ -119,10 +77,10 @@ int main(int argc, char **argv)
 
     while (1) {
         do {
-            send_entropy(random_fd);
+            send_entropy(get_entropy, random_fd);
         } while (ioctl(random_fd, RNDGETENTCNT, &ent_count) == 0 &&
                  ent_count < FILL_WATERMARK);
-        send_entropy(urandom_fd);
+        send_entropy(get_entropy, urandom_fd);
         if (poll(&pfd, 1, MAX_SLEEP) == -1) {
             perror("poll failed");
         }
